fixing a bunch of warnings
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@22660 f5854215-dd47-0410-b2c4-cdd35faa7885

Former-commit-id: 3f12524ba537ff427d8ec6b4ebb52f3db01ab2ec

+++ stacks/geometry/tf/include/tf/message_notifier.h
@@ -183,7 +183,9 @@ public:
   ~MessageNotifier()
   {
     NOTIFIER_DEBUG("Successful Transforms: %llu, Failed Transforms: %llu, Discarded due to age: %llu, Transform messages received: %llu, Messages received: %llu, Total dropped: %llu",
-                        successful_transform_count_, failed_transform_count_, failed_out_the_back_count_, transform_message_count_, incoming_message_count_, dropped_message_count_);
+                   (long long unsigned int)successful_transform_count_, (long long unsigned int)failed_transform_count_, 
+                   (long long unsigned int)failed_out_the_back_count_, (long long unsigned int)transform_message_count_, 
+                   (long long unsigned int)incoming_message_count_, (long long unsigned int)dropped_message_count_);
 
     unsubscribeFromMessage();
